refactor(sharedCheckboxAndRadioStyles): moved to internal/utils
